Fix the issues
Fix the issue of duplicate fuel grade name on product filter of sale summary page.
Fix the issue of BOS ID on filter drop of to time.
Sort the fuel grade name according to client requirement of product wise summary on shift summary page.
Sort the nozzles according to client requirement of pump wise summary on shift summary page.
Remove the average columns on product wise summary on sale summary page.
Fix the time issue on shift summary page.
Fix the sidebar issue on frontend it overlap on navbar when hover on it.